+++ src/generatedTypes/v0.1.0.ts
@@ -6,13 +6,17 @@
  */
 
 /**
- * Semantic versioning of document
+ * Semantic versioning of document.
  */
 export type Version = string;
 /**
  * The code identifying the CLI, UI, service generating the order.
  */
 export type AppCode = string;
+/**
+ * Semantic versioning of document.
+ */
+export type Version1 = string;
 export type ReferrerAddress = string;
 
 /**
@@ -32,7 +36,7 @@ export interface Metadata {
   [k: string]: unknown;
 }
 export interface Referrer {
-  version: Version;
+  version: Version1;
   address: ReferrerAddress;
   [k: string]: unknown;
 }
