+++ src/generatedTypes/v0.3.0.ts
@@ -6,7 +6,7 @@
  */
 
 /**
- * Semantic versioning of document
+ * Semantic versioning of document.
  */
 export type Version = string;
 /**
@@ -17,10 +17,18 @@ export type AppCode = string;
  * Environment from which the order came from
  */
 export type Environment = string;
+/**
+ * Semantic versioning of document.
+ */
+export type Version1 = string;
 export type ReferrerAddress = string;
 export type QuoteId = string;
 export type QuoteSellAmount = string;
 export type QuoteBuyAmount = string;
+/**
+ * Semantic versioning of document.
+ */
+export type Version2 = string;
 
 /**
  * Metadata JSON document for adding information to orders.
@@ -41,7 +49,7 @@ export interface Metadata {
   [k: string]: unknown;
 }
 export interface Referrer {
-  version: Version;
+  version: Version1;
   address: ReferrerAddress;
   [k: string]: unknown;
 }
@@ -49,6 +57,6 @@ export interface Quote {
   id?: QuoteId;
   sellAmount: QuoteSellAmount;
   buyAmount: QuoteBuyAmount;
-  version: Version;
+  version: Version2;
   [k: string]: unknown;
 }
