+++ test/mock.ts
@@ -1,0 +1,11 @@
+export const ADDRESS = '0xb6BAd41ae76A11D10f7b0E664C5007b908bC77C9'
+
+export const MISSING_VERSION_ERROR = [
+  {
+    instancePath: '',
+    keyword: 'required',
+    message: "must have required property 'version'",
+    params: { missingProperty: 'version' },
+    schemaPath: '#/required',
+  },
+]
